New code gen configuration features
-- Added capitalize-names, capitalize-words and word-separator configuration values to
provide additional control over the generation of C++ identifiers from schema names.

-- Added the ability to scope configuration values to all schema contexts beneath a
particular context.  Appending .* to a context specifier instructs the code generator to
apply the associated configuration values to all sub-contexts of the base context, but not
the base context itself. E.g.:

(foo.*):
      capitalize-names: false

applies the given config values to foo.bar, but not to foo itself.